+++ example/src/components/layout/countUp/util.js
@@ -1,0 +1,50 @@
+function decimalAdjust(type, value, exp) {
+    // If the exp is undefined or zero...
+    if (typeof exp === 'undefined' || +exp === 0) {
+        return Math[type](value);
+    }
+    value = +value;
+    exp = +exp;
+    // If the value is not a number or the exp is not an integer...
+    if (isNaN(value) || !(typeof exp === 'number' && exp % 1 === 0)) {
+        return NaN;
+    }
+    // Shift
+    value = value.toString().split('e');
+    value = Math[type](+(value[0] + 'e' + (value[1] ? (+value[1] - exp) : -exp)));
+    // Shift back
+    value = value.toString().split('e');
+    return +(value[0] + 'e' + (value[1] ? (+value[1] + exp) : exp));
+}
+
+if (!Math.round10) {
+    Math.round10 = function(value, exp) {
+        return decimalAdjust('round', value, exp);
+    };
+}
+// Decimal floor
+if (!Math.floor10) {
+    Math.floor10 = function(value, exp) {
+        return decimalAdjust('floor', value, exp);
+    };
+}
+// Decimal ceil
+if (!Math.ceil10) {
+    Math.ceil10 = function(value, exp) {
+        return decimalAdjust('ceil', value, exp);
+    };
+}
+
+const evaluateDecimals = (num) => {
+    let str = String(num);
+    let hasIndex = str.indexOf('.') + 1;
+    if(hasIndex > 0) {
+        let len = str.length;
+        return len - hasIndex;
+    }
+    return 0;
+}
+
+export {
+    evaluateDecimals
+}